pyproject.toml
[tool.poetry]
name = "odtp"
version = "0.2.0"
description = "A tool to deploy and manage open digital twins"
authors = ["caviri <carlosvivarrios@gmail.com>"]
license = "AGPL 3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "2.5.2"
typer = "0.9.0"
pymongo = "3.12.0"
python-dotenv = "^1.0.0"
boto3 = "^1.33.13"
barfi = "^0.7.0"
streamlit = "^1.29.0"
st-pages = "^0.4.5"
streamlit-card = "^1.0.0"
pygwalker = "^0.3.17"
streamlit-aggrid = "^0.3.4.post3"
nicegui = "^1.4.11"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^23.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
odtp = "odtp.cli.main:app"
